% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{age.pyramid}
\alias{age.pyramid}
\title{Plot patient demographics by outcome.}
\usage{
age.pyramid(data, ...)
}
\arguments{
\item{data}{Output of \code{\link{process_data}}. This should be a dataframe which includes columns for age group, sex, and outcome of patients. See 'Details'.}
}
\value{
Bar plot of the age (in intervals of four years) and sex (male/female) of patients, plotted according to clinical outcome (discharge/death/ongoing care).

The columns of \code{data} for age group, sex and outcome should be named "agegp5", "sex" and "outcome" respectively and formatted as follows:
the variable "sex" should be numeric with values 1 and 2 for males and females respectively;
the variable "agegp5" should be a factor with levels 0-4, 5-9, 10-14, ...., 90+;
and the variable "outcome" should be a factor with levels 'discharge', 'censored' and 'death'; in this case, 'censored' patients are those for whom clinical care is ongoing.
}
\description{
Plots the age and sex distribution of patients according to clinical outcome.
}
\details{
#' @export age.pyramid
}
