% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{blood.results.by.age}
\alias{blood.results.by.age}
\title{Box plots for laboratory results within 24 hours of hospital presentation.}
\usage{
blood.results.by.age(data, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
A plot for each laboratory result, showing box plots by age group. On top of each plot, N (the number of individuals
whose records are included in the plot) is printed (this varies between plots due to data completeness).
}
\description{
Plots the following laboratory results by age group: WCC (10^{9}/L), Lymphocytes (10^{9}/L), Neutrophilis (10^{9}/L), Urea (mmol/L), CRP (mg/L),
Prothrombin time (s), APTT (s), Bilirubin (\mu mol/L) and ALT (units/L).
}
