% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{comorb.by.age}
\alias{comorb.by.age}
\title{Plot the prevalence of comorbidities by age group.}
\usage{
comorb.by.age(data, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
A plot for each comorbidity, showing the 95\% binomial confidence interval (represented by a black line)
for the proportion of patients in each age group with that comorbidity. The width of the boxes gives an indication--in relative terms--
of the number of people in each age group for whom information on that comorbidity is available. On top of each plot, N (the number of individuals
whose records are included in the plot) is printed (this varies between plots due to data completeness).
}
\description{
Plots the prevalence of seven comorbidities (asthma, malignancy, HIV, obesity, diabetes mellitus, dementia, and smoking), stratified  by age group.
}
