% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{comorbidity.prevalence.plot}
\alias{comorbidity.prevalence.plot}
\title{Plot distribution of comorbidity prevalance.}
\usage{
comorbidity.prevalence.plot(data, comorbidities, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
Barplot showing the proportion of patients reporting each symptom.
Bars are annotated with a fraction representing the number of patients reporting a comorbidity
over the number of patients for whom presence or absence of that comorbidity was recorded.
}
\description{
Plots the proportion of patients reporting each comorbidity at admission.
}
\details{
Note that the denominators used in the computation of proportions may differ by comorbidity as
information on the presence or absence of some comorbidities may be missing/incomplete for some patients.
}
