% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{generate.report}
\alias{generate.report}
\title{Generate a PDF report from the data}
\usage{
generate.report(patient.data.output, file.name)
}
\arguments{
\item{patient.data.output}{List output from \code{import.and.process.data}}

\item{file.name}{Path to a PDF file for the report}
}
\description{
Generate a PDF report from the data
}
