% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{icu.violin.plot}
\alias{icu.violin.plot}
\title{Plot lengths of hospital stay for patients admitted into Intensive Care Unit (ICU)/High Dependency Unit (HDU).}
\usage{
icu.violin.plot(data, ...)
}
\arguments{
\item{data}{Output of \code{\link{process_data}}.}
}
\value{
Violin plots (with box plots) showing the distribution of the total length of hospital stay for patients who were admitted to ICU/HDU and the
distribution of the lengths of stay within ICU/HDU. The coloured areas of the plot indicate the kernel probability density of the observed data
and the box plots show the median and interquartile range of the lengths of stay.
}
\description{
Plots the distribution of lengths of stay for patients who were admitted to ICU/HDU: the distribution of the total length of hospital stay for this group is plotted,
as well as the length of stay within ICU/HDU.
}
