% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{import.and.process.data}
\alias{import.and.process.data}
\title{Import data from Redcap .csv output for processing}
\usage{
import.and.process.data(
  data.file,
  data.dict.file,
  column.table.file = "required_columns.csv",
  source.name = NA,
  message.out.file = NULL,
  embargo.length = 0,
  ref.date = today(),
  verbose = F
)
}
\arguments{
\item{data.file}{Path of the data file}

\item{data.dict.file}{Path of the data dictionary file}

\item{column.table.file}{Path of the column translation file (default required_columns.csv)}

\item{source.name}{String identifier for the data source (optional)}

\item{message.out.file}{Path of a file to store messages about fields that were overwritten in the data cleaning process. If not given, this file is not generated}

\item{embargo.length}{Length of the data embargo in days; patients enrolled less than this period before \code{ref.date} will be excluded from reports. Default 0.}

\item{ref.date}{Date to be taken as the date of the report. Default is today's date.}

\item{verbose}{Flag for verbose output}
}
\description{
Import data from Redcap .csv output for processing
}
