% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{recruitment.dat.plot}
\alias{recruitment.dat.plot}
\title{Plot cumulative recruitment of patients.}
\usage{
recruitment.dat.plot(data, embargo.limit, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}

\item{embargo.limit}{The cut-off date for inclusion in the report. Patients recruited after \code{embargo.limit} are not considered in the analysis.
Set  \code{embargo.limit} to the date of the report if all patients are to be considered.

this date have not been included.}
}
\value{
Plot showing the cumulative number of patients in the study. One line plots the cumulative of patients for whom follow-up has been recorded
while the other line captures patients for whom follow-up is ongoing. The first dashed black line indicates the \code{embargo.limit}.
The second black line is the cut-off date for the next report, assuming that reports are issued weekly.
}
\description{
Plots the cumulative recruitment of patients, separated by whether follow-up is ongoing or an outcome has been recorded.
}
