% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{signs.by.age}
\alias{signs.by.age}
\title{Box plots for observations at hospital by age group.}
\usage{
signs.by.age(data, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the \code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
A plot for each observation, showing box plots by age group. On top of each plot, N (the number of individuals
whose records are included in the plot) is printed (this varies between plots due to data completeness).
}
\description{
Plots patients' data on five vital signs (respiratory rate, heart rate, systolic blood pressure and temperature) as well as the oxygen staturation in room air (\%), by age group.
Respiratory rate is recorded in breaths per minute, heart rate in beats per minute, systolic blood pressure in mmHg and temperature in degree Celsius.
}
