% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sites.by.country}
\alias{sites.by.country}
\title{Plot the number of sites by country}
\usage{
sites.by.country(data, ...)
}
\arguments{
\item{data}{Output of \code{\link{process_data}}. This should be a dataframe which includes columns for the country and site associated with each patient. See 'Details'.

The columns of \code{data} containing country and site names should be named "Country" and "site.name" respectively.}
}
\value{
Bar plot showing the number of sites per country. Actual counts are printed on top of each bar.
}
\description{
Plots the number of sites by country
}
