% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sx.by.age}
\alias{sx.by.age}
\title{Plot the prevalence of symptoms by age group.}
\usage{
sx.by.age(data, admission.symptoms, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
A plot for each symptom, showing the 95\% binomial confidence interval (represented by a black line)
for the proportion of patients in each age group who presented that symptom. The width of the boxes gives an indication--in relative terms--
of the number of people in each age group for whom information on that symptom is available. On top of each plot, N (the number of individuals
whose records are included in the plot) is printed (this varies between plots due to data completeness).
}
\description{
Plots the prevalence of five symptoms (fever, cough, shortness of breath, confusion, and gastrointestinal symptoms), stratified  by age group.
}
