% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{symptom.heatmap}
\alias{symptom.heatmap}
\title{Plot pairwise symptom prevalance.}
\usage{
symptom.heatmap(data, admission.symptoms, asterisks = vector(), ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
Heatmap showing the proportion of patients for each pairwise combination of symptoms.
}
\description{
Plots a heatmap for prevalance of pairwise combinations of symptoms.
The pairwise prevalence proportions are caculated amongst patients with recorded presence or absence of both symptoms.
}
