% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{symptom.prevalence.plot}
\alias{symptom.prevalence.plot}
\title{Plot distribution of symptom prevalance.}
\usage{
symptom.prevalence.plot(data, admission.symptoms, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}..}
}
\value{
Barplot showing the proportion of patients reporting each symptom.
Bars are annotated with a fraction representing the number of patients presenting with a symptom
over the number of patients for whom presence or absence of that symptom was recorded.
}
\description{
Plots the proportion of patients presenting with each symptom at admission.
}
\details{
Note that the denominators used in the computation of proportions may differ by symptom as symptom information
may be incomplete for some patients.
}
