% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{treatment.upset}
\alias{treatment.upset}
\title{Plot frequency of combinations of treatments.}
\usage{
treatment.upset(data, ...)
}
\arguments{
\item{data}{Output of \code{\link{process_data}}.}
}
\value{
UpSet plot showing the frequency of combinations of the 5 most common treatments.
Filled and empty circles below the x-axis of the plot indicate treatments that were and were not administered respectively.
}
\description{
Plots the distribution of combinations of the 5 most common treatments administered during hospital stay,
across all patients with completed hospital stay and recorded treatment data.
}
