% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{treatment.use.plot}
\alias{treatment.use.plot}
\title{Plot distribution of treatments used.}
\usage{
treatment.use.plot(data, treatments, ...)
}
\arguments{
\item{data}{Output of \code{\link{process_data}}.}
}
\value{
Barplot showing the proportion of patients given each treatment.
Bars are annotated with a fraction representing the number of patients given a treatment
over the number of patients for whom presence or absence of that treatment was recorded.
}
\description{
Plots the proportion of patients given each treatment during clinical care.
}
\details{
Note that the denominators used in the computation of proportions may differ by treatment as
information on treatment given may be missing/incomplete for some patients.
}
