% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{violin.age.func}
\alias{violin.age.func}
\title{Plot lengths of hospital stay by age group}
\usage{
violin.age.func(data, embargo.limit, ...)
}
\arguments{
\item{data}{Output of \code{\link{process_data}}.}
}
\value{
Violin plots (with box plots) showing the distribution of the total length by age group. Age is plotted in  10-year intervals: 0-9, 10-19, \dots, 70+.
The coloured areas of the plot indicate the kernel probability density of the observed data and the box plots show the
median and interquartile range of the lengths of hospital stay for each age group.
}
\description{
Plots the distribution of lengths of stay by age group. Only cases with reported outcomes (i.e. death/discharge) are considered.
}
