% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{generate.report}
\alias{generate.report}
\title{Generate a PDF report from the data.}
\usage{
generate.report(patient.data.output, file.name, site.name)
}
\arguments{
\item{patient.data.output}{List output from \code{import.and.process.data}}

\item{file.name}{Path to a PDF file for the report}

\item{site.name}{Name of the site from which this data is derived}
}
\value{
PDF report containing summaries of the data.
}
\description{
Generates a PDF report containing summaries of the data, including comparison of lengths of hospital stay by sex and age group, outcome by sex,
comorbidity, symptom and treatment distributions and distribution of vital signs on presentation at hospital.
}
