% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{import.and.process.data}
\alias{import.and.process.data}
\title{Import data from Redcap .csv output for processing}
\usage{
import.and.process.data(
  data.file,
  data.dict.file,
  column.table.file = NULL,
  source.name = NA,
  message.out.file = NULL,
  check.early.dates = TRUE,
  embargo.length = 0,
  ref.date = today(),
  yn.field.type = "radio",
  treatment.field.categories = "treatment",
  comorbidity.categories = "comorbidities",
  admission.symptom.field.categories = "admission_signs_and_symptoms",
  overall.switch = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data.file}{Path of the data file}

\item{data.dict.file}{Path of the data dictionary file}

\item{column.table.file}{Path of the column translation file (default required_columns.csv)}

\item{source.name}{String identifier for the data source (optional)}

\item{message.out.file}{Path of a file to store messages about fields that were overwritten in the data cleaning process. If not given, this file is not generated}

\item{embargo.length}{Length of the data embargo in days; patients enrolled less than this period before \code{ref.date} will be excluded from reports. Default 0.}

\item{ref.date}{Date to be taken as the date of the report. Default is today's date.}

\item{yn.field.type}{The "Field Type" in the data dictionary for yes/no questions. Usually "radio" but sometimes "dropdown"}

\item{treatment.field.categories}{The "Form Name" entry or entries (will accept a vector of multiple values) in the data dictionary for treatments.}

\item{comorbidity.categories}{The "Form Name" entry or entries (will accept a vector of multiple values) in the data dictionary for comorbidities.}

\item{admission.symptom.field.categories}{The "Form Name" entry or entries (will accept a vector of multiple values) in the data dictionary for sypmtoms at admission.}

\item{overall.switch}{In some datasets (e.g. RAPID) treatment fields ending "cmyn", "occur" or "prtrt" refer to treatments given on the day of admission, rather than over the entire hospital stay. Overall stay treatments are in columns with "overall_" appending to the start of the name. This option switches these over.}

\item{verbose}{Flag for verbose output}
}
\value{
A list with components: \describe{
\item{unembargoed.data}{Data frame containing patients' records up to \code{ref.date}.}
\item{embargo.limit}{Embargo date; i.e. \code{ref.date} - \code{embargo.length}}
\item{detailed.data}{Data frame containing patients' records up to \code{embargo.limit}}
\item{cst.reference}{Data frame containing name and labels of the symptoms, combordities, and treatments considered.}
}
}
\description{
Import data from Redcap .csv output for processing
}
