% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{outcomes.by.admission.date}
\alias{outcomes.by.admission.date}
\title{Plot weekly admission counts}
\usage{
outcomes.by.admission.date(data, embargo.limit, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}.. This should be a dataframe which includes columns for the date of admission and  outcome for each patient. See `Details'.}
}
\value{
Bar plot showing the number of patients per country and by outcome (discharge/ongoincg care/death). Bars are annotated with counts.
}
\description{
Plots patient numbers and outcomes by epidemiological week (of 2020) of admission (or, for patients infected in hospital, of symptom onset).
}
