% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{outcomes.by.country}
\alias{outcomes.by.country}
\title{Plot the distribution of patients by country and outcome}
\usage{
outcomes.by.country(data, ...)
}
\arguments{
\item{data}{\code{detailed.data}, a component of the output of \code{\link{import.and.process.data}}.. This should be a dataframe which includes columns for the country and outcome associated with each patient. See `Details'.}
}
\value{
Bar plot showing the number of patients per country and by outcome (discharge/ongoing care/death). Actual counts of the total number of patients for each country are printed on top of each bar.
}
\description{
Plots the distribution of patients by country and outcome.
}
